% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICLEN.R
\name{IdentifyFeatures}
\alias{IdentifyFeatures}
\title{SICLEN: Single-cell Clustering based on effective noise reduction through ensenble similarity network}
\usage{
IdentifyFeatures(inData = data, pFeat = 0.1)
}
\arguments{
\item{inData}{single-cell RNA seuqncing data}

\item{pFeat}{Percentage of features to be selected}
}
\value{
List of potential feature genes
}
\description{
This function identifies the potential feature genes
}
\references{
Accurate single-cell clustering through effective noise reduction over ensemble similarity network
}
\author{
Hyundoo Jeong <hdj@inu.ac.kr>
}
