% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICLEN.R
\name{load_obj}
\alias{load_obj}
\title{SICLEN: Single-cell Clustering based on effective noise reduction through ensenble similarity network}
\usage{
load_obj(path)
}
\arguments{
\item{path}{file name to be loaded}
}
\value{
load R objects
}
\description{
This function reduces the noise in the single-cell RNA sequencing through a random walk restart over the ensenble similarity network  the number of clusters in the single-cell RNA sequencing data
}
\references{
Accurate single-cell clustering through effective noise reduction over ensemble similarity network
}
\author{
Hyundoo Jeong <hdj@inu.ac.kr>
}
