% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICLEN.R
\name{make_ensnet}
\alias{make_ensnet}
\title{SICLEN: Single-cell Clustering based on effective noise reduction through ensenble similarity network}
\usage{
make_ensnet(dat_mat, var_genes, K = 30, nPCs = 10, nens = 20)
}
\arguments{
\item{dat_mat}{(M by N) dimensional single-cell RNA seuqncing data}

\item{var_genes}{The set of potential feature genes}

\item{K}{The number of nearest neighbors for KNN networks}

\item{nPCs}{The number of principal components}

\item{nens}{The number of similarity estimates to construct the ensemble similarity network}
}
\value{
Ensemble similarity network based on the different cell-to-cell similarity estimations
}
\description{
This function constructs the ensemble similarity network based on different cell-to-cell similarity estimations
}
\references{
Accurate single-cell clustering through effective noise reduction over ensemble similarity network
}
\author{
Hyundoo Jeong <hdj@inu.ac.kr>
}
