% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICLEN.R
\name{siclen_clust}
\alias{siclen_clust}
\title{SICLEN: Single-cell Clustering based on effective noise reduction through ensenble similarity network}
\usage{
siclen_clust(data, best_nc, Knum = 20)
}
\arguments{
\item{data}{(M x N) dimensional matrix for gene expressions
M: The number of genes
N: The number of cells}

\item{best_nc}{The number of estimated clustes}

\item{Knum}{Initial number of clusters}
}
\description{
SICLEN: Single-cell Clustering based on effective noise reduction through ensenble similarity network
}
\references{
Accurate single-cell clustering through effective noise reduction over ensemble similarity network
}
\author{
Hyundoo Jeong <hdj@inu.ac.kr>
}
