% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICLEN.R
\name{siclen_denoise}
\alias{siclen_denoise}
\title{SICLEN: Single-cell Clustering based on effective noise reduction through ensenble similarity network}
\usage{
siclen_denoise(data, nens = 20, knum = 30, alpha = 0.7, npc = 10, pFeat = 0.05)
}
\arguments{
\item{data}{(M x N) dimensional matrix for gene expressions
M: The number of genes
N: The number of cells}

\item{nens}{The number of similarity estimations}

\item{knum}{The number of nearest neighbors for KNN networks}

\item{alpha}{Restart probability of the random walker}

\item{npc}{The number of principal components}
}
\value{
Noise reduced single-cell RNA sequencing data and learned ensemble similarity network
}
\description{
This function reduces the noise in the single-cell RNA sequencing through a random walk restart over the ensenble similarity network  the number of clusters in the single-cell RNA sequencing data
}
\references{
Accurate single-cell clustering through effective noise reduction over ensemble similarity network
}
\author{
Hyundoo Jeong <hdj@inu.ac.kr>
}
